Move MoveFieldCloserToUsageInspectionTests for QuickFixes into relevant files under RD.QuickFixes
